bldr status command only use status.order field
instead of defaulting to spec.order if the status is not ready.

in the new buildpack world, the spec.order[].group can reference either
a buildpack id + version, or a object reference to a Buildpack or
ClusterBuildpack resource. in order to avoid doing additional lookups in
the case of Buildpack resources, we simply leave the field empty until
the builder fully reconciles

Signed-off-by: Bohan Chen <bohanc@vmware.com>